requirements.txt
Django==1.9
Pillow==3.0.0
argparse==1.2.1
django-bootstrap3==6.2.2
django-ckeditor==5.0.2
wsgiref==0.1.2
